Extract VerifyBlockHandler
