+++ packages/processor/source/verifiers/index.ts
@@ -1,0 +1,1 @@
+export * from "./verify-block-handler";
